% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mime_part.R
\name{mime_part.data.frame}
\alias{mime_part.data.frame}
\title{Create a MIME part from a \code{data.frame}.}
\usage{
\method{mime_part}{data.frame}(x, name = deparse(substitute(x)), ...)
}
\arguments{
\item{x}{A \code{data.frame}.}

\item{name}{Basename of file attachment that is generated.}

\item{...}{Ignored.}
}
\value{
An S3 \code{mime_part} object.
}
\description{
Create a MIME part from a \code{data.frame}.
}

