\name{mime_part.matrix}
\alias{mime_part.matrix}
\title{Create a MIME part from a matrix.}
\usage{\method{mime_part}{matrix}(x, name=deparse(substitute(x)), ...)
}
\description{Create a MIME part from a matrix.}
\value{An S3 \code{mime_part} object}
\arguments{\item{x}{Matrix}
\item{name}{Basename of file attachment that is generated.}
\item{...}{Ignored.}
}

