% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mime_part.R
\name{mime_part}
\alias{mime_part}
\title{Create a MIME part}
\usage{
mime_part(x, name, ...)
}
\arguments{
\item{x}{Object to include}

\item{name}{Name of mime part. Usually the filename of the
attachment as displayed by the e-mail client.}

\item{...}{Possible further arguments for \code{mime_part}
implementations.}
}
\value{
An S3 \code{mime_part} object.
}
\description{
Create a MIME part
}
