% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mime_part.trellis}
\alias{mime_part.trellis}
\title{Creates a MIME part from a trellis plot object}
\usage{
\method{mime_part}{trellis}(x, name = deparse(substitute(x)), device = pdf,
  ...)
}
\arguments{
\item{x}{A \code{trellis} (lattice) object}

\item{name}{Name of attachment (sans .pdf extension).}

\item{device}{Graphics device used to render the plot. Defaults to
\code{pdf}.}

\item{...}{Ignored.}
}
\value{
An S3 \code{mime_part} object.
}
\description{
Writes a PDF file of the plot defined by \code{x} and turns this
PDF file into a file attachment.
}

