% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{sendmail_options}
\alias{sendmail_options}
\alias{sendmailOptions}
\title{Set package specific options.}
\usage{
sendmail_options(...)

sendmailOptions(...)
}
\arguments{
\item{...}{Any options can be defined, using \code{name=value} or
by passing a list of such tagged values.  However, only the ones
below are used in base sendmailR.}
}
\value{
For \code{sendmail_options()}, a list of all set options
sorted by name. For \code{sendmail_options(name)}, a list of length
one containing the set value, or 'NULL' if it is unset.  For uses
setting one or more options, a list with the previous values of
the options changed (returned invisibly).
}
\description{
Specify global sendmail options so that subsequent calls to
\code{sendmail()} do not have to set them in the \code{control}
argument.
}
\details{
List of options:
\itemize{
\item{smtpServer}{SMTP server to contact. This can either be the
  mail server responsible for the destination addresses domain or a
  smarthost provided by your ISP or institution. SMTP AUTH is
  currently unsupported.}
\item{smtpPort}{SMTP port to use. Usually 25 but some institutions
  require the use of the submission service (port 587).}
\item{verbose}{Show detailed information about message
  submission. Useful for debugging.}
}
}
\author{
Olaf Mersmann \email{olafm@datensplitter.net}
}
