\name{mime_part.data.frame}
\alias{mime_part.data.frame}
\title{Create a MIME part from a data...}
\usage{\method{mime_part}{data.frame}(x, name=deparse(substitute(x)), ...)
}
\description{Create a MIME part from a \code{data.frame}.}
\value{An S3 \code{mime_part} object.}
\arguments{\item{x}{A \code{data.frame}.}
\item{name}{Basename of file attachment that is generated.}
\item{...}{Ignored.}
}

