\name{mime_part}
\alias{mime_part}
\title{Create a MIME part...}
\usage{mime_part(x, name, ...)
}
\description{Create a MIME part}
\value{An S3 \code{mime_part} object.}
\arguments{\item{x}{Object to include}
\item{name}{Name of mime part. Usually the filename of the
attachment as displayed by the e-mail client.}
\item{...}{Possible further arguments for \code{mime_part}
implementations.}
}

